Add test_spec parameters for podspecs with different test_spec name.

+++ scripts/code_coverage_report/pod_test_code_coverage_report.sh
@@ -14,27 +14,58 @@
 
 set -ex
 
-SDK="$1"
-platform="$2"
-default_output_path="/Users/runner/${SDK}-${platform}.xcresult"
-output_path="${3:-${default_output_path}}"
+# Loop through arguments and process them
+for arg in "$@"
+do
+  case $arg in
+    --sdk=*)
+      SDK="${arg#*=}"
+      shift # Remove --sdk= from processing
+      ;;
+    --platform=*)
+      PLATFORM="${arg#*=}"
+      shift
+      ;;
+    --test_spec=*)
+      TEST_SPEC="${arg#*=}"
+      shift
+      ;;
+    --output_path=*)
+      OUTPUT_PATH="${arg#*=}"
+      shift
+      ;;
+  esac
+done
+DEFAULT_OUTPUT_PATH="/Users/runner/${SDK}-${PLATFORM}.xcresult"
+DEFAULT_TEST_SPEC="unit"
+OUTPUT_PATH="${OUTPUT_PATH:-${DEFAULT_OUTPUT_PATH}}"
+TEST_SPEC="${TEST_SPEC:-${DEFAULT_TEST_SPEC}}"
+
+[[ -z "$SDK" ]] && { echo "Parameter --sdk should be specified, e.g. --sdk=FirebaseStorage" ; exit 1; }
+[[ -z "$PLATFORM" ]] && { echo "Parameter --platform should be specified, e.g. --platform=ios" ; exit 1; }
+
+echo "SDK: ${SDK}"
+echo "PLATFORM: ${PLATFORM}"
+echo "OUTPUT_PATH: ${OUTPUT_PATH}"
+echo "TEST_SPEC: ${TEST_SPEC}"
+
 if [ -d "/Users/runner/Library/Developer/Xcode/DerivedData" ]; then
-rm -r /Users/runner/Library/Developer/Xcode/DerivedData/*
+  rm -r /Users/runner/Library/Developer/Xcode/DerivedData/*
 fi
 
 # Setup for pod unit tests
 if [ $SDK == "FirebasePerformance" ]; then
   scripts/setup_bundler.sh
-  scripts/third_party/travis/retry.sh scripts/build.sh Performance ${platform} unit
+  scripts/third_party/travis/retry.sh scripts/build.sh Performance ${PLATFORM} unit
 elif [ $SDK == "FirebaseFirestore" ]; then
-  scripts/install_prereqs.sh Firestore ${platform} xcodebuild
-  scripts/third_party/travis/retry.sh scripts/build.sh Firestore ${platform} xcodebuild
+  scripts/install_prereqs.sh Firestore ${PLATFORM} xcodebuild
+  scripts/third_party/travis/retry.sh scripts/build.sh Firestore ${PLATFORM} xcodebuild
 else
   # Run unit tests of pods and put xcresult bundles into output_path, which
   # should be a targeted dir of actions/upload-artifact in workflows.
   # In code coverage workflow, files under output_path will be uploaded to
   # Github Actions.
-  scripts/third_party/travis/retry.sh scripts/pod_lib_lint.rb "${SDK}".podspec --platforms="$(tr '[:upper:]' '[:lower:]'<<<${platform})" --test-specs=unit
+  scripts/third_party/travis/retry.sh scripts/pod_lib_lint.rb "${SDK}".podspec --platforms="$(tr '[:upper:]' '[:lower:]'<<<${PLATFORM})" --test-specs="${TEST_SPEC}"
 fi
 
 find /Users/runner/Library/Developer/Xcode/DerivedData -type d -regex ".*/.*\.xcresult" -execdir cp -R '{}' "${output_path}" \;
